Calculate the gene order score

Calculate the gene order score.

geneOrderScore(
  genesList,
  ids,
  ref,
  k = length(ids),
  max_gap = 1e+07,
  method = c("edit distance", "global alignment score", "spearman correlation"),
  output = c("score", "value matrix")
)

Arguments

genesList

A list of GRanges.

ids

IDs to be plotted. See getGeneCluster.

ref

The reference species.

k

The maximum number of nearest genes used in getGeneCluster.

max_gap

The maximal gaps from the first ID in `ids`.

method

The method to calculate the gene order score. 'edit distance', the gene order score is the mean of edit (Levenshtein) distance of given ids from different species. Gene order score = mean(adist)*k/length(ids). The higher the gene order score, the lower the conservation of the gene order. 'global alignment score', the gene order score is the mean of global alignment score. see pairwiseAlignment. The higher the gene order score, the higher the conservation of the gene order. 'spearman correlation', the score is the mean of Spearman correlation of the genes appearance order.

output

Output values

Value

the score or values matrix such as adist, or alignment scores.

Examples

# example code